Refactored UserSlice
