+++ src/features/User/UserSlice.ts
@@ -1,105 +1,9 @@
-import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
-import axios from 'axios';
+import { createSlice } from '@reduxjs/toolkit';
 import { RootState } from '../../app/store';
-import { API_URL } from '../../constants';
-import { deleteToken, setToken } from '../../helpers/auth/auth';
-
-export type UserCredentials = {
-    email: string,
-    password: string,
-};
-
-type SliceState = {
-    _id: string,
-    displayedName: string,
-    name: string,
-    email: string,
-    isFetching: boolean,
-    isSuccess: boolean,
-    isError: boolean,
-    errorMsg: string,
-}
-
-type UserAttributes = {
-    _id: string,
-    email: string,
-    name: string,
-    displayedName: string,
-}
-
-type LoginResponse = {
-    token: string,
-    user: UserAttributes,
-}
-
-type ErrorResponse = {
-    error: string
-}
-
-type RegisterResponse = {
-    message: string,
-}
-
-export type RegisterCredentials = {
-    email: string,
-    name: string,
-    displayedName: string,
-    password: string,
-};
-
-export const loginUser = createAsyncThunk<
-    LoginResponse,
-    UserCredentials,
-    {
-        rejectValue: ErrorResponse,
-    }
->(
-    'user/login',
-    async (userCredentials, thunkApi) => {
-        try {
-            const response = await axios.post(
-                `${API_URL}/users/login`, {
-                email: userCredentials.email,
-                password: userCredentials.password,
-            }
-            )
-            if (response.status === 200) {
-                setToken(response.data.token);
-                return response.data as LoginResponse;
-            } else {
-                return thunkApi.rejectWithValue(response.data as ErrorResponse);
-            }
-        } catch (error) {
-            console.log("Error", error.response.data);
-            return thunkApi.rejectWithValue(error.response.data as ErrorResponse);
-        }
-    }
-);
-
-export const registerUser = createAsyncThunk<
-    RegisterResponse,
-    RegisterCredentials,
-    {
-        rejectValue: ErrorResponse,
-    }
->(
-    'user/register',
-    async (userCredentials, thunkApi) => {
-        try {
-            const response = await axios.post(
-                `${API_URL}/users/register`, userCredentials
-            )
-            if (response.status === 200) {
-                return response.data as RegisterResponse;
-            } else {
-                return thunkApi.rejectWithValue(response.data as ErrorResponse);
-            }
-        } catch (error) {
-            console.log("Error", error.response.data);
-            return thunkApi.rejectWithValue(error.response.data as ErrorResponse);
-        }
-    }
-);
+import { deleteToken } from '../../helpers/auth/auth';
+import { UNKNOWN_ERROR_MSG } from '../common';
+import { SliceState } from "./types";
+import { loginUser, registerUser } from './thunks';
 
 const initialState: SliceState = {
     _id: '',
@@ -132,6 +36,9 @@ export const userSlice = createSlice({
         }
     },
     extraReducers: (builder) => {
+
+        //login
+
         builder.addCase(loginUser.fulfilled, (state, { payload }) => {
             state._id = payload.user._id;
             state.displayedName = payload.user.displayedName;
@@ -145,12 +52,15 @@ export const userSlice = createSlice({
         builder.addCase(loginUser.rejected, (state, action) => {
             state.isFetching = false;
             state.isError = true;
-            state.errorMsg = action.payload ? action.payload.error : "unknown error occured";
+            state.errorMsg = action.payload ? action.payload.error : UNKNOWN_ERROR_MSG;
         });
 
         builder.addCase(loginUser.pending, (state) => {
             state.isFetching = true;
         });
+
+        //register
+
         builder.addCase(registerUser.fulfilled, (state) => {
             state.isFetching = false;
             state.isSuccess = true;
@@ -160,7 +70,7 @@ export const userSlice = createSlice({
         builder.addCase(registerUser.rejected, (state, action) => {
             state.isFetching = false;
             state.isError = true;
-            state.errorMsg = action.payload ? action.payload.error : "unknown error occured";
+            state.errorMsg = action.payload ? action.payload.error : UNKNOWN_ERROR_MSG;
         });
 
         builder.addCase(registerUser.pending, (state) => {
@@ -171,3 +81,5 @@ export const userSlice = createSlice({
 
 export const userSelector = (state: RootState) => state.user;
 export const { clearState, resetState, logout } = userSlice.actions;
+export {loginUser, registerUser};
+export type {LoginCredentials, RegisterCredentials} from './types';
